requirements.txt
sqlite3==3.45.3
pandas==2.2.2
scikit-learn==1.5.1
pyyaml==6.0.1
joblib==1.4.2